Player movement bugfix & ResetPlane
The HUGE change here is that I realized we need to also reset the player's rotation when they are reset (either from the newly introduced `ResetPosition` or after a roll). With this, I don't think we can really get off the grid with a 1-unit cube. Small cube is still borked though.
This commit also introduces a ResetPlane which when triggered resets the player.